functions/sendNotification/requirements.txt
# https://cloud.google.com/functions/docs/writing/specifying-dependencies-python
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
firebase-admin==4.1.0
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-secret-manager==0.2.0
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
h2==2.6.2
hpack==3.0.0
httplib2==0.17.3
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
msgpack==1.0.0
pathspec==0.8.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
rsa==4.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9